backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
scikit-learn==1.4.2
shap==0.45.0
imbalanced-learn==0.12.2
numpy==1.26.4
python-multipart==0.0.9
reportlab==4.1.0
scipy==1.13.0
google-genai>=1.0.0
